Fix loading custom arrays from different packages
Need to update the package when we resolve types, was a problem with
e.g.:

	type User struct {
		AltEmails mailaddress.List
	}

And then in `mailaddress`:

	type List []Address

It would keep the package name as `model`, so it would resolve to
`model.Address`, instead of `mailaddress.Address`.

Would only happen some of the time, since some times `model.Address`
would get loaded first.