Fix JSON output
Turns out that the omitempty tag behaves different for the JSON and YAML
encoders. The YAML one is more advanced and checks if all fields in a
struct are the zero value, whereas the JSON one simply checks if the
field is the zero value or nil.

So use pointers to avoid adding empty objects to the JSON output.