Don't store embedded structs as context
The Context parameter was used both to signal the location we found the
struct (e.g. response, query params, etc.) and whether or not it's an
embedded struct (which needs to be treated different).

The problem with this is that context information is lost when using
embedded structs in e.g. query params, leading to subtle problems such
as using the wrong struct tags.

This fixes that by adding an IsEmbed parameter.

Fixes #36