Drop $ref
1) Drop `$ref`; you can now use this:

    Path: createRequest

The reason `$ref` was used in the first place was because originally you
could describe parameters inline:

    Path:
        id (int): The foo ID.

So a special "keyword" was needed to distinguish between inline
parameters and references to types.

But suppor for this has long been removed, so it's no longer needed.
It's just overly verbose.

2) Use `{keyword}` instead of `$keyword` for reference directives.

This is more consistent with how keywords are formatted in parameters
(e.g. `{required}`).

---

Both changes will make adding variables (#59) easier, so you can do:

    GET /foo

    Get the foo object. This endpoint is limited to $rateLimit requests
    per second.

    Response body: fooType

It's also simpler syntax, which should make it easier to use for people.

Should be easy to `sed` existing projects to the new syntax:

    find . \! -path './vendor/*' -a -name '*.go' \
        -exec sed -i -e 's/: \$empty/: {empty}/; s/: \$data/: {data}/; s/: \$default/: {default}/; s/: \$ref: /: /g' {} \;
    sed -i 's/: \$ref: /: /g' ./etc/kommentaar.conf