Add Extend paramter
The `Extend` parameter is optional and allows you to extend the generated
endpoint with extra data. For example using Stoplight.io you can specify an
endpoint as private with a non-standard key in the schema. For example:

    Extend: private.yaml

Where `private.yaml` contains the extra key/values for the schema:

    x-private: true

See the endpoint-extend test for full example.