Add tags property to output
Adds the tags property to the OpenAPI root object with all the tags that
occurred in the output sorted. We currently output the tags to the endpoint but
some tools use the top level tags array for the ordering of them.

This supports extra properties where you can add a description so this could be
configuration based at some point, but will take this simple approach unless
somebody else requires it.

See: https://swagger.io/specification/